Add files via upload
In this work we will show concepts about protection of personal data, according to
privacy laws. It will be discussed the importance of a more sophisticated regulation,
because there is a lot of sales of personal data through large corporations, as well as
the existence of alternative ways to remain in secrecy on the internet, addressing
topics such as anonymity in the "Deep Web" network, displaying the operation of the
encryption used. As well as a conclusion about the existence of privacy on the
internet.